feat: объединяет настройки css переменных и свойств Mui
