+++ packages/client/src/components/app-header/AppHeader.module.scss
@@ -55,6 +55,10 @@
       display: block;
       width: 100%;
       padding: 0.5rem 2.5rem;
+
+      &:hover {
+        background: transparent;
+      }
     }
   }
