+++ packages/client/src/styles/index.scss
@@ -4,8 +4,6 @@
 @import "@fontsource/roboto/500.css";
 @import "@fontsource/roboto/700.css";
 @import "variables/global";
-@import "themes/normal";
-@import "themes/dark";
 
 html {
   font-size: 62.5%;
