--- packages/client/src/styles/themes/dark.scss
@@ -1,8 +1,0 @@
-.app.dark {
-  --bg-color: #333;
-  --text-color: #fff;
-  --primary-light-color: #013d74;
-  --primary-main-color: #01386a;
-  --primary-dark-color: #061e3a;
-  --secondary-color: blue;
-}
