--- packages/client/src/styles/themes/normal.scss
@@ -1,6 +1,0 @@
-.app.light {
-  --bg-color: #e8e8ea;
-  --text-color: #333;
-  --primary-color: coral;
-  --secondary-color: blue;
-}
