+++ packages/client/src/theme/ThemeProvider.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo, useState } from "react";
 import { LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext } from "./ThemeContext";
 import { ThemeProvider as ThemeProviderMui, createTheme } from "@mui/material/styles";
-import { muiPallete } from "./muiPallete";
+import { muiPallete, setGlobalStyles } from "./muiPallete";
 
 const defaultTheme = (localStorage.getItem(LOCAL_STORAGE_THEME_KEY) as Theme) || Theme.DARK;
 
@@ -29,6 +29,7 @@ const ThemeProvider: React.FC<React.PropsWithChildren> = ({ children }) => {
 
   return (
     <ThemeProviderMui theme={themeMui}>
+      {setGlobalStyles(themeMui)}
       <ThemeContext.Provider value={defaultProps}>{children}</ThemeContext.Provider>
     </ThemeProviderMui>
   );
